+++ src/Symfony/Component/Security/Core/Encoder/BasePasswordEncoder.php
@@ -21,6 +21,14 @@ abstract class BasePasswordEncoder implements PasswordEncoderInterface
     const MAX_PASSWORD_LENGTH = 4096;
 
     /**
+     * {@inheritdoc}
+     */
+    public function needsRehash(string $encoded): bool
+    {
+        return false;
+    }
+
+    /**
      * Demerges a merge password and salt string.
      *
      * @param string $mergedPasswordSalt The merged password and salt string