+++ src/Symfony/Component/Security/Core/User/User.php
@@ -157,4 +157,9 @@ final class User implements UserInterface, EquatableInterface, AdvancedUserInter
 
         return true;
     }
+
+    public function setPassword(string $password)
+    {
+        $this->password = $password;
+    }
 }